Final Commit
- Fixed bugs
- Added pattern "dd/mm/yy" to Chimpum's code
- Full working association Artifact-Chimpum

+++ src/main/java/acme/forms/AdministratorDashboard.java
@@ -55,7 +55,7 @@ public class AdministratorDashboard {
 	
 	//Individual
 	
-	int										ratioOfArtifactsWithChimpum;
+	double										ratioOfArtifactsWithChimpum;
 	
 	Map<String,Double>						averageBudgetOfChimpumsGroupedByCurrency;
 	Map<String,Double>						deviationBudgetOfChimpumsGroupedByCurrency;
